+++ magpylib/_lib/classes/base.py
@@ -36,35 +36,37 @@ Define base classes here on which the magnetic source objects are built on
 
 '''
 
-
-######### Type hint definitions ########
-# These aren't type hints, but look good 
-# in Spyder IDE. Pycharm recognizes it.
+# tool-tip / intellisense helpers -----------------------------------------------
+# Class initialization is done purely by kwargs. While some # of these can be 
+# set to zero by default other MUST be given to make any sense 
+# (e.g. magnetization). To improve tool tips and intellisense we inizilize them
+# with names, e.g. mag=(Mx, My, Mz). This looks good, but it requires that
+# these names are pre-initialzed:
 Auto = 0 # Maximum cores, for multicore
         # function. if 0 find max.
 numpyArray = 0
 constant = None
 Mx=My=Mz=0.0 # Zero Moment
-#######################################
 
-# %% IMPORTS
+
+# -------------------------------------------------------------------------------
 from numpy import array, float64, pi, isnan, array
 from magpylib._lib.mathLibPrivate import Qmult, Qconj, getRotQuat, arccosSTABLE, fastSum3D, fastNorm3D
 from magpylib._lib.utility import checkDimensions
-from magpylib._lib.classes.fieldsampler import FieldSampler
 import sys
 
 
-# %% FUNDAMENTAL CLASS - RCS (RELATIVE COORDINATE SYSTEM)
+# -------------------------------------------------------------------------------
+# FUNDAMENTAL CLASS - RCS (RELATIVE COORDINATE SYSTEM)
 #       - initiates position, orientation
 #       - adds moveBY, rotateBy
 
 class RCS:
     """
-    FUNDAMENTAL CLASS - RCS (RELATIVE COORDINATE SYSTEM)
+    base class RCS(RELATIVE COORDINATE SYSTEM)
 
-    initiates position, orientation
-    - adds moveBY, rotateBy
+    initiates position, orientation (angle, axis)
+    adds methods setPosition, move, setOrientation, rotate
     """
 
     def __init__(self, position, angle, axis):
@@ -262,7 +264,51 @@ class RCS:
         self.position = array(Vnew[1:])+anchor
 
 
-# %% HOMOGENEOUS MAGNETIZATION CLASS
+#------------------------------------------------------------------------------
+class FieldSampler:
+    """
+    Field Sampler Class
+    
+    This class initiates the getB method and is inherited by all source objects.
+    The main reason this is centralized here is that the docstring of the
+    getB method is inherited by all sources.
+    """
+
+    def getB(self, pos):
+        """
+        This method returns the magnetic field vector generated by the source
+        at the argument position `pos` in units of [mT]
+
+        Parameters
+        ----------
+        pos : vec3 [mm] Position or list of Positions where magnetic field
+            should be determined.
+
+
+        Returns
+        -------
+        magnetic field vector : arr3 [mT] Magnetic field at the argument
+            position `pos` generated by the source in units of [mT].
+        """
+        # Return a list of vec3 results
+        # This method will be overriden by the classes that inherit it.
+        # Throw a warning and return 0s if it somehow isn't.
+        # Note: Collection() has its own docstring
+        # for getB since it inherits nothing.
+        import warnings
+        warnings.warn(
+            "called getB method is not implemented in this class,"
+            "returning [0,0,0]", RuntimeWarning)
+        return [0, 0, 0]
+
+
+
+#------------------------------------------------------------------------------
+#------------------------------------------------------------------------------
+
+
+
+# subclass for HOMOGENEOUS MAGNETS --------------------------------------
 #       - initiates magnetization
 
 class HomoMag(RCS, FieldSampler):
@@ -280,7 +326,7 @@ class HomoMag(RCS, FieldSampler):
             self.magnetization) == 3), "Bad mag input, invalid vector dimension"
 
 
-# %% LINE CURRENT CLASS
+# subclass for LINE CURRENTS ---------------------------------------------
 #       - initiates current
 
 class LineCurrent(RCS, FieldSampler):
@@ -297,6 +343,8 @@ class LineCurrent(RCS, FieldSampler):
             sys.exit('Bad current input')
 
 
+# subclass for MOMENTS ----------------------------------------------------
+#       - initiates nothing
 class MagMoment(RCS, FieldSampler):
 
     def __init__(self, moment=(Mx, My, Mz),
